programs add and delete fix tt-47
