+++ tertuliatalk_frontend/pages/_app.tsx
@@ -31,7 +31,6 @@ const navItems: NavItems = [
   { title: 'Kayıt ol', href: '/signup' },
 
   { title: 'Profil', href: '/profile' },
-  { title: 'Giriş yap', href: '/', outlined: true },
 ];
 
 const TinaCMS = dynamic(() => import('tinacms'), { ssr: false });
